Lower priority of HIDTest task
Commit @583 uncomments the HIDTest task creation.  HIDTest task is configured to run at
priority 4 (== tskIDLE_PRIORITY + 4) which is higher priority than the "System" task at
priority 3 (== tskIDLE_PRIORITY + 3).

The HIDTest task never blocks so it prevents the system task (and any other task of priority
less than 4) from ever running.

This commit does not fix the root problem of HIDTest never blocking, but rather lowers
its priority to be equal to the system task so that they share the CPU.  This is a
temporary workaround.

git-svn-id: svn://svn.openpilot.org/OpenPilot/trunk@586 ebee16cc-31ac-478f-84a7-5cbb03baadba