gcs serial: fix serial port telemetry on Linux
On Linux, port.portName is the simple name of the serial port
like 'ttyS0'.  The physical name of the device is '/dev/ttyS0'.
This value is stored in port.physName.  This bug doesn't get
in the way on Windows since 'COM1' is both the simple name as
well as the physical name.  On Linux, the distinction is
important.

git-svn-id: svn://svn.openpilot.org/OpenPilot/trunk@710 ebee16cc-31ac-478f-84a7-5cbb03baadba