demo_slam: Added computation of state perturbation in robot. Also added operation with camera (decomented lines). Adjusted velocity perturbation levels. Reconfigured camera to 15Hz.
The systems works, but:

there is a segFault if predicted features fall out of the image.
the update is not very robust and tracking is inaccurate. This I think is because of the missed frames.
We can try to compute dt at each time and update the pertubation levels accordingly.