Default location and model changed for model view. This means adding a ModelView plugin should just work without any configuration.
There is still an issue with this plugin where it will crash if no model is selected, this is not a work-around and that issue needs resolving. However this does mean that it will not meet that issue by default (as is the case now) as it will default to the Easystar.  

git-svn-id: svn://svn.openpilot.org/OpenPilot/trunk@964 ebee16cc-31ac-478f-84a7-5cbb03baadba