Fix the IMU noises to make inertial slam work!
We must not divide the noise density by the acquisition frequency,
because the std dev does not depend on the acquisition frequency,
and because the full computation did not have consistent units.
We multiply by dt from P_continuous_time to P because we integrate
the value, not because of the acquisition frequency.

There is a 1.4 factor for gyros because the experimental noise (0.006)
is slightly larger than computed noise (0.0055). For acceleros experimental
noise (0.009) is compatible with computed noise (0.0011).

Moreover xsens has increased the accelero noise density from 0.001
to 0.002 m/s2/sqrt(Hz) between the datasheet releases of 2005 and 2009.