Only update if innovation is significant wrt measurement uncertainty.
Large updates are causing inconsistency because of linearization errors,
but too numerous updates are also causing inconsistency,
so we should avoid to do not significant updates.
An update is not significant if there are large odds that it is
only measurement noise and that there is not much information.

When the camera is not moving at all, the landmarks are converging anyway
quite fast because of this, at very unconsistent positions of course,
so that when the camera moves it cannot recover them.

Some work needs to be done yet to prevent search ellipses from growing
too much and integrate it better with the whole management, but this was
for first evaluation purpose.

Unfortunately it doesn't seem to improve much the situation, even if
it is still working correctly with less computations.
The feature is disabled for now.