gcs uploader: add support for "Safe Boot" to the GCS uploader
When halted in the bootloader or while rescuing a board, the
user can press the "Safe Boot" button in the uploader gadget
to force the FW to boot with a default hwsettings configuration.

The default conditions of the hwsettings uavo will disable all
optional modules, disable all serial port config, and ensure that
the board can communicate via the USB HID telemetry interface.

Once booted in this mode, a user can easily reconfigure the
hwsettings uavo through the config GUI and save the fixed
settings to the board to be used on the next reboot.  No need
to wipe all settings just to recover from a non-functional
HW config.

NOTE: The GCS needs to grow some very clear visual clues to
      indicate when the board has booted in safe mode.  The
      firmware helpfully raises a (new) critical alarm called
      BootFault whenever it boots in safe mode.