qt_sdk: upgrade to SDK v1.2 (QT v4.8.0)
Qt 4.8+ is now required to build the GCS.

This version of the SDK installer has dropped support
for the --installdir option so users will need to manually
specify the path for installation as:
  <top>/tools/qtsdk-v1.2
in order for the Makefile to pick up this toolchain automatically.

This update drops support for Ubuntu 9.10 (Karmic) due to an
incompatibility with the glibc that ships with that release.

Support should still be possible with Ubuntu 10.04 (Lucid) and
newer.