Fix RTH mode.  Now works in simulation when GCSReceiver goes invalid.
There is a small issue that when it is in RTH failsafe mode and you regain
signal and your throttle is low you will disarm due to the timeout.  That
means you'd suddenly get it back and plumet ouf of the sky.  If you are
still holding throttle (which most people might do?) you will have control.

The timeout needs to probably start when you regain signal.