makefile: change QT SDK install dir so we can support multiple installs
QT SDK version was changed but the install dir wasn't changed
to match.  This makes it impossible to keep both versions
installed at the same time which is essential in order to work
on/build old and new branches.

This update gives the new SDK version a unique install dir so
it can coexist with older versions.