AndroidGCS Telemetry: Sending an update could interrupt a pending transaction
We allow sending objects that don't create a transaction while waiting for one
to complete.  However, this was overriding the pending transaction information
(when there wasn't a new transaction) and masking the transaction completed
message.

Note: I think this logical error can occur on the GCS side and might explain the
"transaction completed when not expected" messages.