[OP-685] Hide the debug window in the GCS by default
Currently GCS prints in the debug window too many messages confusing users.
Until it is fixed (so only important messages are printed) the debug
windows is removed from default GCS configurations. It still is available
as a widget and can be added to a GCS workspace.

It is better than hide it using options checkbox (as suggested in OP-685)
and keep empty window frame.