build-info.txt: minor text fix
This template will be used to comment auto builds created with Bamboo.
Currently it is emulated by Bamboo build agents using build scripts,
but should be generated using Makefile. As soon as majority of branches
will be updated to this next, those scripts will be removed and this
template used.