f4 usb: ensure correct interface numbers for BL and FW
The interface numbers differ for the HID interface between
the BL and FW.  This is required to keep windows happy.

In HID+CDC, the CDC interfaces must come first with HID last
which makes it #2.  In HID only mode, we can't use #2 for HID
since windows doesn't like the first/only interface to be
anything other than #0.

Conflicts:
	flight/targets/board_hw_defs/discoveryf4/board_hw_defs.c
	flight/targets/board_hw_defs/flyingf4/board_hw_defs.c
	flight/targets/board_hw_defs/freedom/board_hw_defs.c
	flight/targets/board_hw_defs/quanton/board_hw_defs.c