OP-719: add linux x64 version of qt-4.8.4
It seems we must ship GCS with all Qt libs like we do for Windows.
On Ubuntu x64 the GCS built with this crashes using system libs.

But since it's possible to build GCS using natively installed
Qt SDK, it is not a problem for now. Something is better than
nothing, and two ways of building are better then just one.

+review OPReview-441