Fix weird error of running python script with extra --verbose option as a part of script name
This error of building any EF targets was found on Windows if used with
make V=1 verbose output option. The '--verbose' python script option was
used by the Windows make as a part of python script file name with
'file not found' error. Now it works. Still can't get the cause of that
error, though...