OP-893: do not include *.opfw firmware files into package
Now GCS has all firmware embedded and available through the Vehicle Setup
Wizard or Auto-Update function. Hence it is not necessary to include
another copies of the same files. But the code to do this is kept because
later we will include sim targets as firmware. Linux version does include
it already, OSX and Windows should follow.