For controls and Dialogs, use Copydir directive as all qml files are needed and this is the "easy way", the other solution is to list every single qml file by hand, these will likely be added to in upcomming releases of Qt and cause maintaince.
The downside is three debug DLLs get copied also, if required I can expand
copydata.pro to delete them after the copy if required.

As we are making some changes to the build system, this at least enables
working builds of QtQuick2, if before packaging we do not have a better
solution, I will make sure the extra debug DLLs are removed.