OP-969 - Fixed unnecessary resizing of input config wizard
Instead of dynamicly adding/removing widgets to the
single wizard page, created all the widgets ahead
and added them to QStackedWidget.

QStackedWidget keeps the size of the biggest child,
so re-layouting is not necessary.

Moved wizard UI to the separate .ui form
as original input.ui and/or implementation
file would be to overloaded otherwise.